+++ packages/astro/src/@types/view-transitions.ts
@@ -1,0 +1,25 @@
+export interface TransitionAnimation {
+	name: string; // The name of the keyframe
+	delay?: number | string;
+	duration?: number | string;
+	easing?: string;
+	fillMode?: string;
+	direction?: string;
+}
+
+export interface TransitionAnimationPair {
+	old: TransitionAnimation | TransitionAnimation[];
+	new: TransitionAnimation | TransitionAnimation[];
+}
+
+export interface TransitionDirectionalAnimations {
+	forwards: TransitionAnimationPair;
+	backwards: TransitionAnimationPair;
+}
+
+export type TransitionAnimationValue =
+	| 'initial'
+	| 'slide'
+	| 'fade'
+	| 'none'
+	| TransitionDirectionalAnimations;
